user_service/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
bcrypt==3.2.0
PyJWT==2.2.0
python-dotenv==0.19.1
alembic==1.7.4
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
kafka-python==2.1.5